Add files via upload
I recommend flipping the model 180 degrees so that the "correct" facing logo is on the plate. If the model is too large, resizing it is an option.

This model will require multicolor printing (i.e. Bambu's AMS). Coloring the model in your slicer should be easy.

Disclaimer: I am not an expert at 3D modelling and the model is no where near optimally created. For the use case of a simple keychain it works perfectly fine, though. Some filament waste will be had with the multicolor printing.